requirements.txt
# Driver Drowsiness Detection - Dépendances
# ==========================================

# Deep Learning
tensorflow>=2.12.0
keras>=2.12.0

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=9.5.0

# Machine Learning
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Data Manipulation
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# UI / Web App
streamlit>=1.26.0
flask>=2.3.0

# Audio / Alerts
pygame>=2.5.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
imutils>=0.5.4

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Pour déploiement sur GPU
# tensorflow-gpu>=2.12.0  # Décommenter si GPU disponible
